docshell/test/test_bug123696.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=123696
-->
<head>
  <title>Test for Bug 123696</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=123696">Mozilla Bug 123696</a>
<p id="display">
  <iframe src="bug123696-subframe.html"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 123696 **/
SimpleTest.waitForExplicitFinish();

function finishTest() {
  is(window.frames[0].frames[0].document.documentElement.textContent,
     "change2", "Reload should have reloaded correctly!");
  SimpleTest.finish();
}

function doReload() {
  window.frames[0].frameElement.onload = finishTest;
  window.frames[0].location.reload();
}

addLoadEvent(function() {
  window.frames[0].frames[0].frameElement.onload = doReload;
  window.frames[0].frames[0].frameElement.src = "javascript:parent.change2()";
});



</script>
</pre>
</body>
</html>
